Refactor: split symbolic_executor.py into focused modules (#99)
Split the 3022-line monolith into 9 modules following the dependency graph
from issue #99. The old module retains re-exports so every import in the
existing `__all__` still resolves — no caller changes required.

Modules (dependency order, leaves first):
  - poly.py            Poly, Monomial, _norm_coeff, _mono_mul, _mono_str
  - modpoly.py         ModPoly (ℤ/2³², issue #78)
  - symbolic_types.py  IndicatorPoly (#76), RationalPoly + SymbolicRemainder
                       (#75), relation constants + helpers
  - closed_form.py     ClosedForm, ProductForm (issue #89)
  - bitvec.py          BitVec AST + _apply_bitop (issue #77)
  - guarded.py         Guard, GuardedPoly, type unions
                       (RationalStackValue, SymbolicIntAst)
  - arithmetic_ops.py  ArithmeticOps hook (issue #68 S3) + DEFAULT
  - symbolic_executor.py  (slim 436 lines) opcode constants,
                       SymbolicResult, exceptions, run_symbolic,
                       + re-exports for backward compat
  - forking_executor.py  run_forking, _Path, ForkingResult,
                       recurrence solver, collapse_report,
                       guarded_to_mermaid (issue #70)

Type unions live in guarded.py (first module that transitively imports every
symbolic-value type). Opcode-classification constants (_POLY_OPS,
_BIN_OP_RELATION, _BITVEC_OPCODES, ...) stay in symbolic_executor.py since
run_symbolic uses them and forking_executor imports them from there.

Verification:
  - Import surface preserved: every name in the old `__all__` importable
    from symbolic_executor, including private names that sibling modules
    (ff_symbolic, etc.) import (_apply_bitop, _REL_SYMBOL, _relation_holds,
    _trunc_div, _trunc_rem, MASK32, ...).
  - Test results bit-for-bit identical to main on test_symbolic_executor,
    test_closed_form, test_symbolic_programs_catalog, test_poly_compiler,
    test_ff_symbolic: 25 failed, 347 passed (pre-existing failures
    unrelated to this refactor).

Per issue #99 goal: agent sessions drop from 50K-token file reads to
<5K tokens per relevant module.
